permissions

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -2,6 +2,8 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
+const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const jwtHelper = userId => jwt.sign({ userId }, process.env.APP_SECRET);
 
@@ -12,18 +14,24 @@ const expireTime = {
 
 const Mutations = {
 	async createItem(parent, args, ctx, info) {
-		// TODO: Check if they are logged in
+		if (!ctx.request.userId) {
+			throw new Error('You must be logged in to do that');
+		}
 
 		const item = await ctx.db.mutation.createItem(
 			{
 				data: {
+					// creates a relationship between item and user
+					user: {
+						connect: {
+							id: ctx.request.userId
+						}
+					},
 					...args
 				}
 			},
 			info
 		);
-
-		console.log(item);
 
 		return item;
 	},
@@ -43,13 +51,25 @@ const Mutations = {
 			info
 		);
 	},
-	async deleteItem(parent, args, ctx, info) {
-		const where = { id: args.id };
-		// 1. find the item
-		const item = await ctx.db.query.item({ where }, `{ id title}`);
-		// 2. Check if they own that item, or have the permissions
-		return ctx.db.mutation.deleteItem({ where }, info);
-	},
+
+  async deleteItem(parent, args, ctx, info) {
+    const where = { id: args.id };
+    // 1. find the item
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
+    // 2. Check if they own that item, or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
+
+    // 3. Delete it!
+    return ctx.db.mutation.deleteItem({ where }, info);
+  },
+
 	async signup(parent, args, ctx, info) {
 		try {
 			args.email = args.email.toLowerCase();
@@ -106,6 +126,16 @@ const Mutations = {
 			where: { email },
 			data: { resetToken, resetTokenExpiry }
 		});
+
+		await transport.sendMail({
+			from: 'wes@wesbos.com',
+			to: user.email,
+			subject: 'Your Password Reset Token',
+			html: makeANiceEmail(`Your Password Reset Token is here!
+      \n\n
+      <a href="${process.env.FRONTEND_URL}/reset?resetToken=${resetToken}">Click Here to Reset</a>`)
+		});
+
 		return { message: 'Thanks' };
 	},
 
@@ -136,6 +166,38 @@ const Mutations = {
 		ctx.response.cookie('token', token, expireTime);
 
 		return updatedUser;
+	},
+
+	async updatePermissions(parent, args, ctx, info) {
+		// 1. Check if they are logged in
+		if (!ctx.request.userId) {
+			throw new Error('You must be logged in!');
+		}
+		// 2. Query the current user
+		const currentUser = await ctx.db.query.user(
+			{
+				where: {
+					id: ctx.request.userId
+				}
+			},
+			info
+		);
+		// 3. Check if they have permissions to do this
+		hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+		// 4. Update the permissions
+		return ctx.db.mutation.updateUser(
+			{
+				data: {
+					permissions: {
+						set: args.permissions
+					}
+				},
+				where: {
+					id: args.userId
+				}
+			},
+			info
+		);
 	}
 };
